requirements.txt
fastapi[all]
firebase_admin
python-dotenv
